+++ scripts/upgrade.sh
@@ -14,12 +14,30 @@
 #  limitations under the License.
 #  ========================================================================    
 
-pem_file=$1
+usage() {
+    cat <<EOM
+    Usage: scripts/$(basename $0) param1
 
+           param1: <pem_file> (e.g. conf/pdb-key.pem)
+
+EOM
+   exit -1;
+}
+
+[ -z $1 ] && { usage; }
+
+pem_file=$1
 user=ubuntu
 ip_len_valid=3
 pdb_dir=$PDB_INSTALL
-echo "To strip shared libraries..."
+testSSHTimeout=3
+
+if [ ! -f ${pem_file} ]; then
+    echo "Pem file '$pem_file' not found, make sure the path and file name are correct!"
+    exit -1;
+fi
+
+echo "To strip shared libraries, this may take some time..."
 strip libraries/*.so
 echo "stripped all shared libraries!"
 
@@ -35,27 +53,30 @@ else
   PDB_SSH_OPTS="-i ${pem_file} $PDB_SSH_OPTS"
 fi
 
-if [ "$PLINY_HOME" = "" ]; then
-  echo "We do not have pliny dependency."
-else
-  rm -rf bin
-  mkdir bin
-  scripts/syncWithPliny.sh
-fi
+while read line
+do
+   [[ $line == *#* ]] && continue # skips commented lines
+   [[ ! -z "${line// }" ]] && arr[i++]=$line # include only non-empty lines
+done < $PDB_HOME/conf/serverlist
 
-arr=($(awk '{print $0}' $PDB_HOME/conf/serverlist))
 length=${#arr[@]}
-echo "There are $length servers"
+echo "There are $length servers defined in $PDB_HOME/conf/serverlist"
+
 for (( i=0 ; i<=$length ; i++ ))
 do
-        ip_addr=${arr[i]}
-        if [ ${#ip_addr} -gt "$ip_len_valid" ]
-        then
-                echo -e "\n+++++++++++ install server: $ip_addr"
-                ssh $PDB_SSH_OPTS $user@$ip_addr "rm $pdb_dir/log.out; rm $pdb_dir/logs/*"
-                scp $PDB_SSH_OPTS -r $PDB_HOME/bin/pdb-worker $user@$ip_addr:$pdb_dir/bin/ 
-                scp $PDB_SSH_OPTS -r $PDB_HOME/scripts/cleanupNode.sh $PDB_HOME/scripts/startWorker.sh $PDB_HOME/scripts/stopWorker.sh $PDB_HOME/scripts/checkProcess.sh $user@$ip_addr:$pdb_dir/scripts/
-        fi
+   ip_addr=${arr[i]}
+   if [ ${#ip_addr} -gt "$ip_len_valid" ]
+   then
+      # checks that ssh to a node is possible, times out after 3 seconds
+      nc -zw$testSSHTimeout ${ip_addr} 22
+      if [ $? -eq 0 ]
+      then
+         echo -e "\n+++++++++++ install server: $ip_addr"
+         ssh $PDB_SSH_OPTS $user@$ip_addr "rm $pdb_dir/log.out; rm $pdb_dir/logs/*"
+         scp $PDB_SSH_OPTS -r $PDB_HOME/bin/pdb-worker $user@$ip_addr:$pdb_dir/bin/ 
+         scp $PDB_SSH_OPTS -r $PDB_HOME/scripts/cleanupNode.sh $PDB_HOME/scripts/startWorker.sh $PDB_HOME/scripts/stopWorker.sh $PDB_HOME/scripts/checkProcess.sh $user@$ip_addr:$pdb_dir/scripts/
+      fi
+   fi
 done
